Extract ActionCollection out of ResourceReporter
This creates a more general-purpose ActionCollection as a list of all
the actions that the chef-client run takes, including all sub-resources
in nested custom resources.

The ResourceReporter now only has the concerns of what it does to
talk to the network and some information it gathers about the overall
chef-client run.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>